Style proposal iteration
This is a draft of the style for the agenda and speakers

The speakers is still not closed and the talks page may be improved
for accessiblity and better readability.

The agenda currently looks the user's timezone.

We want to add a timezone selector to change the timezones everywhere.
The calendar.js file is already prepared for that change when we make it
happen in the style pages.